Updated objects to index items by integer not long.

+++ app/src/model/task.go
@@ -2,8 +2,8 @@ package model
 
 // Task is an Object for holding a task.
 type Task struct {
-	ID                   int64  `json:"id"`
-	UserID               int64  `json:"userId"`
+	ID                   int32  `json:"id"`
+	UserID               int32  `json:"userId"`
 	Memo                 string `json:"memo"`
 	RepeatIntervalInDays int64  `json:"repeatIntervalInDays"`
 	TaskLength           int64  `json:"taskLength"`
