+++ app/src/model/user.go
@@ -2,7 +2,7 @@ package model
 
 // User is an account that owns Tasks.
 type User struct {
-	ID       int64  `json:"id"`
+	ID       int32  `json:"id"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
 }
